+++ src/Frontend/Dfe.Complete/Services/ProjectPermissionService.cs
@@ -11,6 +11,7 @@ namespace Dfe.Complete.Services
         bool UserCanComplete(ProjectDto project, ClaimsPrincipal user);
         bool UserCanEdit(ProjectDto project, ClaimsPrincipal user);
         bool UserCanDaoRevocation(ProjectDto project, ClaimsPrincipal user);
+        bool UserIsAdmin(ProjectDto project, ClaimsPrincipal user);
     }
 
     public class ProjectPermissionService : IProjectPermissionService
@@ -22,6 +23,9 @@ namespace Dfe.Complete.Services
             UserIsServiceSupport(user) || (ProjectIsActive(project) && UserIsAssignee(project, user));
         public bool UserCanDaoRevocation(ProjectDto project, ClaimsPrincipal user) =>
             ProjectIsActive(project) && (UserIsServiceSupport(user) || UserIsAssignee(project, user)) && project.DirectiveAcademyOrder == true;
+
+        public bool UserIsAdmin(ProjectDto project, ClaimsPrincipal user) =>
+            UserIsServiceSupport(user);
 
         private static bool ProjectIsActive(ProjectDto project) => project.State == ProjectState.Active;
         private static bool UserIsAssignee(ProjectDto project, ClaimsPrincipal user) => project.AssignedToId?.Value == user.GetUserId().Value;