+++ src/Tests/Dfe.Complete.Tests/Models/ExternalContact/ExternalContactAddEditPageModelTests.cs
@@ -20,6 +20,7 @@ namespace Dfe.Complete.Tests.Models.ExternalContact
     using MediatR;
     using Microsoft.Extensions.Logging;
     using Moq;
+    using System;
     using System.Threading.Tasks;
     using Xunit;
 
@@ -64,6 +65,7 @@ namespace Dfe.Complete.Tests.Models.ExternalContact
             ProjectId projectId = fixture.Create<ProjectId>();
             ContactId contactId = fixture.Create<ContactId>();
             string urn = "123";
+            string incomingTrustUkprn = "101483";
             string localAuthorityCode = fixture.Create<string>();
 
             var mockTrustDto = fixture.Build<TrustDto>().With(x => x.Name, input).Create();
@@ -82,7 +84,10 @@ namespace Dfe.Complete.Tests.Models.ExternalContact
                   .With(t => t.Id, projectId)
                   .With(t => t.Type, ProjectType.Transfer)
                   .With(t => t.EstablishmentName, input)
-                  .With(t => t.Urn, new Urn(Convert.ToInt32(urn)))
+                  .With(t => t.Urn, new Urn(Convert.ToInt32(urn)))                         
+                  .With(t => t.IncomingTrustUkprn, new Ukprn(Convert.ToInt32(incomingTrustUkprn)))
+                  .With(t => t.NewTrustReferenceNumber, string.Empty)
+                  .With(t => t.NewTrustName, string.Empty)
                   .Create())
 
                .With(t => t.ExternalContactInput, fixture.Build<OtherExternalContactInputModel>()